BIT/requirements.txt
snowflake-connector-python
uuid
